consolidated similar member stores, fixed order of member types, and integrated new "isAdmin" functionality from backend
